fix: Let FollowUpItemResponse.avg_interval_days carry a fractional EWMA interval
Optional[int] rejected any non-whole Decimal, so a cycle with a real EWMA
interval (e.g. 19.3318) raised a 500 from get_follow_ups instead of a
response. Widened the field to Optional[Decimal] and added a regression
test that flows a fractional interval through the orchestrator end to end.

+++ app/schemas/sale.py
@@ -174,7 +174,7 @@ class FollowUpItemResponse(BaseModel):
     """A single product that a customer is expected to need."""
     product_id: uuid.UUID
     product_name: str
-    avg_interval_days: Optional[int] = None
+    avg_interval_days: Optional[Decimal] = None
     last_purchase_date: Optional[date] = None
     last_quantity: Decimal = Decimal("0")
     estimated_next_purchase: Optional[date] = None
